+++ app/src/main/java/org/simple/clinic/contactpatient/ContactPatientUiRenderer.kt
@@ -3,6 +3,7 @@ package org.simple.clinic.contactpatient
 import org.simple.clinic.contactpatient.UiMode.CallPatient
 import org.simple.clinic.contactpatient.UiMode.SetAppointmentReminder
 import org.simple.clinic.home.overdue.OverdueAppointment
+import org.simple.clinic.home.overdue.OverduePatientAddress
 import org.simple.clinic.mobius.ViewRenderer
 import org.simple.clinic.overdue.PotentialAppointmentDate
 import org.simple.clinic.overdue.TimeToAppointment
@@ -26,8 +27,16 @@ class ContactPatientUiRenderer(
 
     ui.hideProgress()
     when (model.uiMode) {
-      CallPatient -> renderCallPatientView(model)
+      CallPatient -> renderCallPatientViewBasedOnFeatureFlag(model)
       SetAppointmentReminder -> renderSetAppointmentReminderView(model)
+    }
+  }
+
+  private fun renderCallPatientViewBasedOnFeatureFlag(model: ContactPatientModel) {
+    if (model.overdueListChangesFeatureEnabled) {
+      renderCallPatientView(model)
+    } else {
+      renderCallPatientView_Old(model)
     }
   }
 
@@ -38,9 +47,9 @@ class ContactPatientUiRenderer(
     ui.switchToSetAppointmentReminderView()
   }
 
-  private fun renderCallPatientView(model: ContactPatientModel) {
+  private fun renderCallPatientView_Old(model: ContactPatientModel) {
     if (model.hasLoadedPatientProfile) {
-      renderPatientProfile(model.patientProfile!!)
+      renderPatientProfile_Old(model.patientProfile!!)
     }
 
     if (model.hasLoadedAppointment) {
@@ -48,12 +57,40 @@ class ContactPatientUiRenderer(
     }
 
     if (model.secureCallingFeatureEnabled) {
+      ui.showSecureCallUi_Old()
+    } else {
+      ui.hideSecureCallUi_Old()
+    }
+
+    ui.switchToCallPatientView_Old()
+  }
+
+  private fun renderCallPatientView(model: ContactPatientModel) {
+    if (model.hasLoadedPatientProfile && model.hasLoadedAppointment) {
+      renderPatientProfile(model.patientProfile!!, model.appointment!!)
+    }
+
+    if (model.patientProfileHasPhoneNumber && model.hasLoadedAppointment) {
+      ui.showPatientWithPhoneNumberUi()
+      ui.hidePatientWithNoPhoneNumberUi()
+      ui.setResultOfCallLabelText()
+
+      loadSecureCallingUi(model)
+    } else {
+      ui.showPatientWithNoPhoneNumberUi()
+      ui.hidePatientWithPhoneNumberUi()
+      ui.setResultLabelText()
+    }
+
+    ui.switchToCallPatientView()
+  }
+
+  private fun loadSecureCallingUi(model: ContactPatientModel) {
+    if (model.secureCallingFeatureEnabled) {
       ui.showSecureCallUi()
     } else {
       ui.hideSecureCallUi()
     }
-
-    ui.switchToCallPatientView()
   }
 
   private fun toggleStateOfReminderDateSteppers(model: ContactPatientModel) {
@@ -100,20 +137,50 @@ class ContactPatientUiRenderer(
 
   private fun toggleCallResultSection(appointment: ParcelableOptional<OverdueAppointment>) {
     if (appointment.isEmpty()) {
-      ui.hideCallResultSection()
+      ui.hideCallResultSection_Old()
     } else {
-      ui.showCallResultSection()
+      ui.showCallResultSection_Old()
     }
   }
 
-  private fun renderPatientProfile(patientProfile: PatientProfile) {
+  private fun renderPatientProfile_Old(
+      patientProfile: PatientProfile
+  ) {
     val patientAge = DateOfBirth.fromPatient(patientProfile.patient, clock).estimateAge(clock)
 
-    ui.renderPatientDetails(
+    ui.renderPatientDetails_Old(
         name = patientProfile.patient.fullName,
         gender = patientProfile.patient.gender,
         age = patientAge,
         phoneNumber = patientProfile.phoneNumbers.first().number
     )
+  }
+
+  private fun renderPatientProfile(
+      patientProfile: PatientProfile,
+      appointment: ParcelableOptional<OverdueAppointment>
+  ) {
+    val patientAge = DateOfBirth.fromPatient(patientProfile.patient, clock).estimateAge(clock)
+
+    ui.renderPatientDetails(PatientDetails(
+        name = patientProfile.patient.fullName,
+        gender = patientProfile.patient.gender,
+        age = patientAge,
+        phoneNumber = patientProfile.phoneNumbers.firstOrNull()?.number,
+        patientAddress = patientAddressText(appointment.get().patientAddress),
+        registeredFacility = appointment.get().appointmentFacilityName.orEmpty(),
+        diagnosedWithDiabetes = appointment.get().diagnosedWithDiabetes,
+        diagnosedWithHypertension = appointment.get().diagnosedWithHypertension,
+        lastVisited = appointment.get().patientLastSeen
+    ))
+  }
+
+  private fun patientAddressText(patientAddress: OverduePatientAddress) = when {
+    !patientAddress.streetAddress.isNullOrBlank() && !patientAddress.colonyOrVillage.isNullOrBlank() -> {
+      "${patientAddress.streetAddress}, ${patientAddress.colonyOrVillage}"
+    }
+    !patientAddress.streetAddress.isNullOrBlank() -> patientAddress.streetAddress
+    !patientAddress.colonyOrVillage.isNullOrBlank() -> patientAddress.colonyOrVillage
+    else -> "${patientAddress.district}, ${patientAddress.state}"
   }
 }