+++ app/src/main/java/org/simple/clinic/contactpatient/views/CallPatientView.kt
@@ -11,6 +11,9 @@ import org.simple.clinic.contactpatient.PatientDetails
 import org.simple.clinic.databinding.ContactpatientCallpatientBinding
 import org.simple.clinic.medicalhistory.Answer
 import org.simple.clinic.patient.displayLetterRes
+import org.simple.clinic.util.UserClock
+import org.simple.clinic.util.toLocalDateAtZone
+import java.time.format.DateTimeFormatter
 
 private typealias AgreedToVisitClicked = () -> Unit
 private typealias RemindToCallLaterClicked = () -> Unit
@@ -88,6 +91,12 @@ class CallPatientView(
       patientWithoutPhoneNumberGroup.visibility = if (field) View.VISIBLE else View.GONE
     }
 
+  var setResultOfCallLabelText: String = resources.getString(R.string.contactpatient_result_of_call)
+    set(value) {
+      field = value
+      resultOfCallLabelTextView.text = field
+    }
+
   var agreedToVisitClicked: AgreedToVisitClicked? = null
 
   var remindToCallLaterClicked: RemindToCallLaterClicked? = null
@@ -118,7 +127,8 @@ class CallPatientView(
 
   fun renderPatientDetails(
       patientDetails: PatientDetails,
-      lastVisited: String
+      dateTimeFormatter: DateTimeFormatter,
+      userClock: UserClock
   ) {
     val genderLetter = resources.getString(patientDetails.gender.displayLetterRes)
     val diagnosis = diagnosisText(patientDetails.diagnosedWithDiabetes, patientDetails.diagnosedWithHypertension)
@@ -128,7 +138,7 @@ class CallPatientView(
     patientAddressTextView.text = patientDetails.patientAddress
     registeredFacilityTextView.text = patientDetails.registeredFacility
     diagnosisTextView.text = diagnosis
-    lastVisitedTextView.text = lastVisited
+    lastVisitedTextView.text = dateTimeFormatter.format(patientDetails.lastVisited.toLocalDateAtZone(userClock.zone))
   }
 
   private fun diagnosisText(diagnosedWithDiabetes: Answer?, diagnosedWithHypertension: Answer?): String {