+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Enqueue `UpdateRemoteConfigWorker` on app cold starts
 - Use `SyncConfigType` qualifier for daily and frequent sync configs
 - Bump Kotlin to v1.5.21
+- Updated `ContactPatientBottomSheet` behaviour for supporting patients with and without phone number
 
 ### Features
 - [In Progress: 06 Jul 2021] Add support for finding a patient online from ID scan